Resolved issue of numerically truncated k-points on ultra dense band paths

+++ src/defs/kpnts_interpolation_mesh.py
@@ -551,15 +551,14 @@ def get_path(ibrav,alat,cell,dk,b_vectors,band_path,special_points):
 
         path_file+="%s %s\n"%(point1,numK)
 
-        a0 = np.linspace(p1[0],p2[0],numK,endpoint=False).astype(np.float16)
-        a1 = np.linspace(p1[1],p2[1],numK,endpoint=False).astype(np.float16)
-        a2 = np.linspace(p1[2],p2[2],numK,endpoint=False).astype(np.float16)
+        a0 = np.linspace(p1[0],p2[0],numK,endpoint=False)
+        a1 = np.linspace(p1[1],p2[1],numK,endpoint=False)
+        a2 = np.linspace(p1[2],p2[2],numK,endpoint=False)
 
         kx = np.concatenate((kx,a0))
         ky = np.concatenate((ky,a1))
         kz = np.concatenate((kz,a2))
         
-        ### Why print? Is this just a test? print("%s %s %s %s %s %s"%(p1[0],p1[0],p1[0],numK,"#",point1))
       except Exception as e:
         print('Exception in kpnts_interpolation_mesh', e)
